results/rebuttal/run-20231111-034650_LLMatic-zero_shot_donnemartin-system-design-oop-bookclub_0_20-runs_log/bookclub/17/LLMatic/meeting.py
class Meeting:
	def __init__(self):
		self.id = None
		self.title = None
		self.description = None
		self.time = None
		self.attendees = None

	def schedule_meeting(self, title, description, time, attendees):
		# Mock database interaction
		meetings_db = {}
		meeting_id = len(meetings_db) + 1
		meetings_db[meeting_id] = {'title': title, 'description': description, 'time': time, 'attendees': attendees}
		return meeting_id

	def get_meeting(self, id):
		# Mock database interaction
		meetings_db = {}
		return meetings_db.get(id, 'Meeting not found')

	def update_meeting(self, id, title, description, time, attendees):
		# Mock database interaction
		meetings_db = {}
		if id in meetings_db:
			meetings_db[id] = {'title': title, 'description': description, 'time': time, 'attendees': attendees}
			return 'Meeting updated'
		return 'Meeting not found'

	def delete_meeting(self, id):
		# Mock database interaction
		meetings_db = {}
		if id in meetings_db:
			del meetings_db[id]
			return 'Meeting deleted'
		return 'Meeting not found'
